refactor(observability): 删除测试专用的旧对话目录适配器
